+++ src/all_storages/delete_any/custom_delete_any.rs
@@ -5,17 +5,17 @@ use hashbrown::hash_set::HashSet;
 
 /// Trait used as a bound for AllStorages::delete_any.
 pub trait CustomDeleteAny {
-    fn delete_any(&mut self, ids: &mut HashSet<EntityId>, current: u32);
+    fn delete_any(&mut self, ids: &mut HashSet<EntityId>, current: u64);
 }
 
 impl CustomDeleteAny for () {
     #[inline]
-    fn delete_any(&mut self, _: &mut HashSet<EntityId>, _current: u32) {}
+    fn delete_any(&mut self, _: &mut HashSet<EntityId>, _current: u64) {}
 }
 
 impl<T: Component> CustomDeleteAny for SparseSet<T, T::Tracking> {
     #[inline]
-    fn delete_any(&mut self, ids: &mut HashSet<EntityId>, current: u32) {
+    fn delete_any(&mut self, ids: &mut HashSet<EntityId>, current: u64) {
         ids.extend(&self.dense);
         self.private_clear(current);
     }
